+++ android/src/main/kotlin/com/example/clevertap_signedcall_flutter/Constants.kt
@@ -14,6 +14,7 @@ object Constants {
     const val KEY_ALLOW_PERSIST_SOCKET_CONNECTION = "allowPersistSocketConnection"
     const val KEY_PROMPT_RECEIVER_READ_PHONE_STATE_PERMISSION = "promptReceiverReadPhoneStatePermission"
     const val KEY_OVERRIDE_DEFAULT_BRANDING = "overrideDefaultBranding"
+    const val KEY_PROMPT_PUSH_PRIMER = "promptPushPrimer"
     const val KEY_MISSED_CALL_ACTIONS = "missedCallActions"
     const val KEY_BG_COLOR = "bgColor"
     const val KEY_FONT_COLOR = "fontColor"