+++ android/src/main/kotlin/com/example/clevertap_signedcall_flutter/SCMethodCall.kt
@@ -4,6 +4,7 @@ object SCMethodCall {
     //Flutter to Platform
     const val INIT = "init"
     const val CALL = "call"
+    const val DISCONNECT_SIGNALLING_SOCKET = "disconnectSignallingSocket"
     const val LOGOUT = "logout"
     const val HANG_UP_CALL = "hangUpCall"
     const val LOGGING = "logging"