+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## CHANGE LOG
 
+### Version 0.0.2 (February 21, 2022)
+* Supports Signed Call Android SDK v0.0.2 and Signed Call iOS SDK v0.0.2.
+* Adds Push Primer support for the [Android 13 notification runtime permission](https://developer.android.com/develop/ui/views/notifications/notification-permission).
+* Adds new public API `disconnectSignallingSocket()` in order to close the Signalling socket connection.
+* Fixes VoIP call screen distortion in iOS.
+
 ### Version 0.0.1 (December 20, 2022)
 * Initial Release.
-* Supports Signed Call Android SDK v0.0.1
-* Supports Signed Call iOS SDK v0.0.1
+* Supports Signed Call Android SDK v0.0.1 and Signed Call iOS SDK v0.0.1.